TRANSFER-804: Fix - preprod transfer kinesis -> CH
Issue (Transfer-804): Kubernetes POD restarting due to a panic in the code caused by accessing a nil pointer upon triggering transfer from kinesis to clickhouse in pre-prod.

Issue cause: A missing check to verify that the pointer is not null before attempting to de-reference it,

Solution: Verify the type of pointer before calling the function.

For more context about the issue: a DLQ RawToTableImpl with an instance variable "dlq" was set to nil and the DLQ pointer was attempting calling the method Do using that nil pointer.
a72067d728e82ac7153c03ae6ce3adbfbf852d96