TRANSFER-735: Test connection succeeds despite non-existent target cluster
Fix about getting *Connection established successfully* message after clicking **Test Connection** button during Transfer creation. The problem is that the target endpoint of this transfer does not exist - or actually the cluster that contains the endpoint does not exist.

Now this should prevent creation in this case, and the test connection should fail at that step.
248ecd83846081b064c557c282ef1f32ca78eff0